% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromXtoR_mixed.R
\name{fromXtoRMixed}
\alias{fromXtoRMixed}
\title{Get an estimate of latent correlation matrix}
\usage{
fromXtoRMixed(X, type = NULL, use.nearPD = TRUE, deriv = FALSE)
}
\arguments{
\item{X}{A numeric data matrix (n by p), n is the sample size and p is the number of variables. Binary variables must be coded as 0, 1. Ordinal variables with l ordered categories must be coded as 0, 1, 2, ... , (l-1).}

\item{type}{A vector of length p containing type of variables in \code{X}, must be in the format "cont", "trunc", "ord", "bin". If not supplied, our function tries to guess the data-type from the observations}

\item{use.nearPD}{A logical value indicating whether to use \link[Matrix]{nearPD} or not when the resulting latent correlation matrix is not positive definite.}

\item{deriv}{A logical value indicating whether to calculate point-wise derivatives of the bridging functions useful for calculating the asymptotic confidence interval of latent correlations}
}
\value{
\code{fromXtoRmixed} returns
\itemize{
\item{hatR: }{the estimated latent correlation matrix}
\item{hatRprime: }{if (deriv = TRUE), returns the pointwise derivative of the bridging function with respect to latent correlation}
}
}
\description{
Estimation of latent correlation matrix from observed data of mixed types (continuous/truncated/ordinal/binary) based on the Semiparametric Gaussian copula model.
}
\examples{
require(SGCTools)
# generate data and estimate latent correlation matrix
n <- 100
p.seq <- 5
k <- 1
p <- p.seq[k]
# Generate random correlation matrix
S <- clusterGeneration::rcorrmatrix(p)
# Generate latent normal
L <- mvtnorm::rmvnorm(n,sigma = S)
# Generate latent non-paranormal (monotone transformations)
Z <- cbind(L[,1]^3, exp(L[,2]), L[,3], L[,4], L[,5])
# Generate observed X : C, B, O, O, T
X <- cbind(Z[,1],Z[,2] >= 1, as.numeric(cut(Z[,3],breaks=c(-Inf,-0.2,0.1,0.3,Inf)))-1 ,as.numeric(cut(Z[,4],breaks=c(-Inf,-0.3,0.05,0.2,Inf)))-1,ifelse(Z[,5]>0,Z[,5],0))
# Calculate Kendall's Tau of observed data
K = Kendall_mixed(X)
# Calculate bridged estimates of latent correlation
R=fromXtoRMixed(X,use.nearPD=TRUE)

# Calculate predictions of latent normal
Lhat <-getLatentPreds(X, lat.cov = R$hatR)
}
\references{
Dey ., Zipunnikov V. (2022) "Semiparametric Gaussian Copula Regression modeling for Mixed Data Types (SGCRM)" \url{https://arxiv.org/abs/2205.06868}

Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" \url{doi:10.1111/rssb.12168}.

Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" \url{doi:10.1093/biomet/asaa007}.

Yoon G., Mueller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" \url{https://arxiv.org/abs/2006.13875}.
}
