% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgclm.R
\name{sgclm}
\alias{sgclm}
\title{Perform latent regression using SGC}
\usage{
sgclm(formula, type = NULL, data)
}
\arguments{
\item{formula}{A regression formula specifying the name of the outcome and covariates in the data frame}

\item{type}{A vector of length p containing type of variables in \code{X}, must be in the format "cont", "trunc", "ord", "bin". If not supplied, our function tries to guess the data-type from the observations}

\item{data}{A numeric data frame (n by p), n is the sample size and p is the number of variables. Binary variables must be coded as 0, 1. Ordinal variables with l ordered categories must be coded as 0, 1, 2, ... , (l-1).}
}
\value{
\code{sgclm} returns a data frame containing the latent regression coefficients, asymptotic standard error, p-values and significance indicator.
}
\description{
Estimate of latent regression coefficient based on outcome and covariates of mixed types (continuous/truncated/ordinal/binary) based on the Semiparametric Gaussian copula model.
}
\examples{
require(SGCTools)
# generate data and estimate latent correlation matrix
n <- 100
p.seq <- 5
k <- 1
p <- p.seq[k]
# Generate random correlation matrix
S <- clusterGeneration::rcorrmatrix(p)
# Generate latent normal
L <- mvtnorm::rmvnorm(n,sigma = S)
# Generate latent non-paranormal (monotone transformations)
Z <- cbind(L[,1]^3, exp(L[,2]), L[,3], L[,4], L[,5])
# Generate observed X : C, B, O, O, T
X <- data.frame(Z[,1],Z[,2] >= 1, as.numeric(cut(Z[,3],breaks=c(-Inf,-0.2,0.1,0.3,Inf)))-1 ,as.numeric(cut(Z[,4],breaks=c(-Inf,-0.3,0.05,0.2,Inf)))-1,ifelse(Z[,5]>0,Z[,5],0))
names(X) <- c("x1","y1","x2","x3","x4")
# Calculate Kendall's Tau of observed data
K = Kendall_mixed(X)
# Calculate bridged estimates of latent correlation
R=fromXtoRMixed(X,use.nearPD=TRUE)

# Principal component analysis using latent correlation estimated using SGC
pc_sgc = princomp(R$hatR,cor=TRUE)

# Perform a latent SGC regression using the second column of X (binary) as an outcome
lm1 = sgclm(y1~ x1+x2 +x4, data=data.frame(X))
lm1

# Calculate predictions of latent normal
Lhat <-getLatentPreds(X, lat.cov = R$hatR)
}
\references{
Dey ., Zipunnikov V. (2022) "Semiparametric Gaussian Copula Regression modeling for Mixed Data Types (SGCRM)" \url{https://arxiv.org/abs/2205.06868}

Fan J., Liu H., Ning Y. and Zou H. (2017) "High dimensional semiparametric latent graphicalmodel for mixed data" \url{doi:10.1111/rssb.12168}.

Yoon G., Carroll R.J. and Gaynanova I. (2020) "Sparse semiparametric canonical correlation analysis for data of mixed types" \url{doi:10.1093/biomet/asaa007}.

Yoon G., Mueller C.L., Gaynanova I. (2020) "Fast computation of latent correlations" \url{https://arxiv.org/abs/2006.13875}.
}
